STEP1: ParcelInformation에서 비용관련 CostInformation 책임 분리
